updated code with automation
